requirements.txt
google-genai
elevenlabs
syncsdk
python-dotenv
srt
numpy
soundfile
librosa
pydub
requests
httpx
pytest
